requirements.txt
Flask==2.3.2
pyharmonics==1.4.3
openai==0.27.2
python-dotenv==1.0.1
gunicorn==20.1.0